LayoutTests/imported/w3c/web-platform-tests/digital-credentials/default-permissions-policy.https.sub-expected.txt



PASS Permissions-Policy is by default 'self' for get().
FAIL Permissions-Policy is by default 'self' for create(). assert_unreached: Should have rejected: undefined Reached unreachable code
